update route favorites

+++ routes/favorites.js
@@ -10,13 +10,14 @@ router.post("/user/favorites/add", isAuthenticated, async (req, res) => {
   try {
     // Check if the game doesn't already exist in the database
     const favorites = await Favorites.findOne({
-      id: req.fields.id,
+      name: req.fields.name,
     });
 
     // If the favorite game doesn't exist in database
     if (!favorites) {
       const newFavorites = new Favorites({
-        id: req.fields.id,
+        // Link the data from the model Favorites to the body of the function addToFavorites
+        owner: req.fields.userId,
         name: req.fields.name,
         image: req.fields.image,
       });
@@ -25,7 +26,7 @@ router.post("/user/favorites/add", isAuthenticated, async (req, res) => {
       // Answer to the client
 
       res.status(200).json({
-        id: newFavorites.id,
+        owner: newFavorites.owner,
         name: newFavorites.name,
         image: newFavorites.image,
       });
@@ -39,31 +40,34 @@ router.post("/user/favorites/add", isAuthenticated, async (req, res) => {
   }
 });
 
-router.delete(
-  "/user/favorites/delete/:id",
-  isAuthenticated,
-  async (req, res) => {
-    try {
-      const id = req.params.id;
-      const resultToDelete = await Favorites.findById(id);
-
-      await resultToDelete.delete();
-
-      res.status(200).json("Favorite deleted succesfully");
-    } catch (error) {
-      res.status(400).json({ message: error.message });
-    }
-  }
-);
-
-router.get("/user/favorites", isAuthenticated, async (req, res) => {
+router.post("/user/favorites/delete", isAuthenticated, async (req, res) => {
   try {
-    const favorites = await Favorites.find()
-      .populate({
-        path: "owner",
-        select: "account",
-      })
-      .select("name image");
+    // Check if the game doesn't already exist in the database
+    const favorites = await Favorites.findOne({
+      name: req.fields.name,
+    });
+    // If the favorite game doesn't exist in database
+    if (!favorites) {
+      await resultToDelete.delete();
+    }
+    res.status(200).json("Favorite deleted succesfully");
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+});
+
+router.get("/user/favorites/:id", isAuthenticated, async (req, res) => {
+  console.log("favorites");
+  try {
+    const userId = req.params.id;
+    console.log(userId);
+    const favorites = await Favorites.find({ owner: userId });
+    // .populate({
+    //   path: "owner",
+    //   select: "account",
+    // })
+    // .select("name image");
+    console.log(favorites);
 
     res.status(200).json({ favorites: favorites });
   } catch (error) {
